Refactor version handling to improve code organization
