+++ workers/manager.ts
@@ -1,13 +1,8 @@
 import fs from "fs";
 import { appendFile } from "fs/promises";
 import path from "path";
-import {
-  formatBytes,
-  generateEncryptionKeyHex,
-  getAutoVersions,
-  sleep,
-  VersionList,
-} from "@helpers/client";
+import { formatBytes, generateEncryptionKeyHex, sleep } from "@helpers/client";
+import { getAutoVersions, VersionList } from "@workers/versions";
 import { type Client, type Group, type XmtpEnv } from "@xmtp/node-sdk";
 import { generatePrivateKey, privateKeyToAccount } from "viem/accounts";
 import { installationThreshold, WorkerClient, type typeofStream } from "./main";
